refactor: change class type name and filename
